requirements.txt
bcrypt==3.1.7
cffi==1.14.0
click==7.1.1
fastapi==0.53.2
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
passlib==1.7.2
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
pyotp==2.3.0
python-multipart==0.0.5
six==1.14.0
starlette==0.13.2
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
requests==2.22.0
authlib==0.14.1
pyopenssl==19.1.0